Refactored nwis packages to put in table specific packages.
