+++ src/main/java/gov/acwi/wqp/etl/nwis/monitoringLocation/NwisMonitoringLocation.java
@@ -1,4 +1,4 @@
-package gov.acwi.wqp.etl.nwis;
+package gov.acwi.wqp.etl.nwis.monitoringLocation;
 
 import java.math.BigDecimal;
