+++ src/main/java/gov/acwi/wqp/etl/nwis/monitoringLocation/NwisMonitoringLocationRowMapper.java
@@ -1,4 +1,4 @@
-package gov.acwi.wqp.etl.nwis;
+package gov.acwi.wqp.etl.nwis.monitoringLocation;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -8,39 +8,39 @@ import org.springframework.jdbc.core.RowMapper;
 
 public class NwisMonitoringLocationRowMapper implements RowMapper<NwisMonitoringLocation> {
 	
-	public static final String SITE_ID_COLUMN_NAME = "site_id";
-	public static final String AGENCY_CD_COLUMN_NAME = "agency_cd";
-	public static final String SITE_NO_COLUMN_NAME = "site_no";
-	public static final String ORGANIZATION_ID_COLUMN_NAME = "organization_id";
-	public static final String PRIMARY_SITE_TYPE_COLUMN_NAME = "primary_site_type";
-	public static final String CALCULATED_HUC_12_COLUMN_NAME = "calculated_huc_12";
-	public static final String HUC_CD_COLUMN_NAME = "huc_cd";
-	public static final String COUNTRY_CD_COLUMN_NAME = "country_cd";
-	public static final String STATE_CD_COLUMN_NAME = "state_cd";
-	public static final String COUNTY_CD_COLUMN_NAME = "county_cd";
-	public static final String GEOM_COLUMN_NAME = "geom";
-	public static final String STATION_NM_COLUMN_NAME = "station_nm";
-	public static final String ORGANIZATION_NAME_COLUMN_NAME = "organization_name";
-	public static final String STATION_TYPE_NAME_COLUMN_NAME = "station_type_name";
-	public static final String DEC_LAT_VA_COLUMN_NAME = "dec_lat_va";
-	public static final String DEC_LONG_VA_COLUMN_NAME = "dec_long_va";
-	public static final String MAP_SCALE_FC_COLUMN_NAME = "map_scale_fc";
-	public static final String LAT_LONG_METHOD_DESCRIPTION_COLUMN_NAME = "lat_long_method_description";
-	public static final String DEC_COORD_DATUM_CD_COLUMN_NAME = "dec_coord_datum_cd";
-	public static final String ALT_DATUM_CD_COLUMN_NAME = "alt_datum_cd";
-	public static final String ALT_VA_COLUMN_NAME = "alt_va";
-	public static final String ALT_ACY_VA_COLUMN_NAME = "alt_acy_va";
-	public static final String ALTITUDE_METHOD_DESCRIPTION_COLUMN_NAME = "altitude_method_description";
-	public static final String DRAIN_AREA_VA_COLUMN_NAME = "drain_area_va";
-	public static final String CONTRIB_DRAIN_AREA_VA_COLUMN_NAME = "contrib_drain_area_va";
-	public static final String LAT_LONG_ACCURACY_COLUMN_NAME = "lat_long_accuracy";
-	public static final String LAT_LONG_ACCURACY_UNIT_COLUMN_NAME = "lat_long_accuracy_unit";
-	public static final String NAT_AQFR_NAME_COLUMN_NAME = "nat_aqfr_name";
-	public static final String AQFR_NM_COLUMN_NAME = "aqfr_nm";
-	public static final String AQUIFER_TYPE_DESCRIPTION_COLUMN_NAME = "aquifer_type_description";
-	public static final String CONSTRUCTION_DT_COLUMN_NAME = "construction_dt";
-	public static final String WELL_DEPTH_VA_COLUMN_NAME = "well_depth_va";
-	public static final String HOLE_DEPTH_VA_COLUMN_NAME = "hole_depth_va";
+	private static final String SITE_ID_COLUMN_NAME = "site_id";
+	private static final String AGENCY_CD_COLUMN_NAME = "agency_cd";
+	private static final String SITE_NO_COLUMN_NAME = "site_no";
+	private static final String ORGANIZATION_ID_COLUMN_NAME = "organization_id";
+	private static final String PRIMARY_SITE_TYPE_COLUMN_NAME = "primary_site_type";
+	private static final String CALCULATED_HUC_12_COLUMN_NAME = "calculated_huc_12";
+	private static final String HUC_CD_COLUMN_NAME = "huc_cd";
+	private static final String COUNTRY_CD_COLUMN_NAME = "country_cd";
+	private static final String STATE_CD_COLUMN_NAME = "state_cd";
+	private static final String COUNTY_CD_COLUMN_NAME = "county_cd";
+	private static final String GEOM_COLUMN_NAME = "geom";
+	private static final String STATION_NM_COLUMN_NAME = "station_nm";
+	private static final String ORGANIZATION_NAME_COLUMN_NAME = "organization_name";
+	private static final String STATION_TYPE_NAME_COLUMN_NAME = "station_type_name";
+	private static final String DEC_LAT_VA_COLUMN_NAME = "dec_lat_va";
+	private static final String DEC_LONG_VA_COLUMN_NAME = "dec_long_va";
+	private static final String MAP_SCALE_FC_COLUMN_NAME = "map_scale_fc";
+	private static final String LAT_LONG_METHOD_DESCRIPTION_COLUMN_NAME = "lat_long_method_description";
+	private static final String DEC_COORD_DATUM_CD_COLUMN_NAME = "dec_coord_datum_cd";
+	private static final String ALT_DATUM_CD_COLUMN_NAME = "alt_datum_cd";
+	private static final String ALT_VA_COLUMN_NAME = "alt_va";
+	private static final String ALT_ACY_VA_COLUMN_NAME = "alt_acy_va";
+	private static final String ALTITUDE_METHOD_DESCRIPTION_COLUMN_NAME = "altitude_method_description";
+	private static final String DRAIN_AREA_VA_COLUMN_NAME = "drain_area_va";
+	private static final String CONTRIB_DRAIN_AREA_VA_COLUMN_NAME = "contrib_drain_area_va";
+	private static final String LAT_LONG_ACCURACY_COLUMN_NAME = "lat_long_accuracy";
+	private static final String LAT_LONG_ACCURACY_UNIT_COLUMN_NAME = "lat_long_accuracy_unit";
+	private static final String NAT_AQFR_NAME_COLUMN_NAME = "nat_aqfr_name";
+	private static final String AQFR_NM_COLUMN_NAME = "aqfr_nm";
+	private static final String AQUIFER_TYPE_DESCRIPTION_COLUMN_NAME = "aquifer_type_description";
+	private static final String CONSTRUCTION_DT_COLUMN_NAME = "construction_dt";
+	private static final String WELL_DEPTH_VA_COLUMN_NAME = "well_depth_va";
+	private static final String HOLE_DEPTH_VA_COLUMN_NAME = "hole_depth_va";
 
 	@Override
 	public NwisMonitoringLocation mapRow(ResultSet rs, int rowNum) throws SQLException {
