+++ src/main/java/gov/acwi/wqp/etl/nwis/nwisDistrictCdsByHost/NwisDistrictCdsByHostRowMapper.java
@@ -1,4 +1,4 @@
-package gov.acwi.wqp.etl.nwis;
+package gov.acwi.wqp.etl.nwis.nwisDistrictCdsByHost;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -8,11 +8,11 @@ import org.springframework.jdbc.core.RowMapper;
 
 public class NwisDistrictCdsByHostRowMapper implements RowMapper<NwisDistrictCdsByHost> {
 	
-	public static final String HOST_NAME_COLUMN_NAME = "host_name";
-	public static final String DISTRICT_CD_COLUMN_NAME = "district_cd";
-	public static final String STATE_NAME_COLUMN_NAME = "state_name";
-	public static final String STATE_POSTAL_CD_COLUMN_NAME = "state_postal_cd";
-	public static final String LAST_UPD_DATE_COLUMN_NAME = "last_upd_date";
+	private static final String HOST_NAME_COLUMN_NAME = "host_name";
+	private static final String DISTRICT_CD_COLUMN_NAME = "district_cd";
+	private static final String STATE_NAME_COLUMN_NAME = "state_name";
+	private static final String STATE_POSTAL_CD_COLUMN_NAME = "state_postal_cd";
+	private static final String LAST_UPD_DATE_COLUMN_NAME = "last_upd_date";
 	
 	@Override
 	public NwisDistrictCdsByHost mapRow(ResultSet rs, int rowNum) throws SQLException {
