+++ src/main/java/gov/acwi/wqp/etl/nwis/result/NwisResult.java
@@ -1,4 +1,4 @@
-package gov.acwi.wqp.etl.nwis;
+package gov.acwi.wqp.etl.nwis.result;
 
 import java.time.LocalDate;
