+++ src/main/java/gov/acwi/wqp/etl/nwis/result/NwisResultRowMapper.java
@@ -1,4 +1,4 @@
-package gov.acwi.wqp.etl.nwis;
+package gov.acwi.wqp.etl.nwis.result;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -9,82 +9,82 @@ import org.springframework.jdbc.core.RowMapper;
 
 public class NwisResultRowMapper implements RowMapper<NwisResult> {
 	
-	public static final String STATION_ID_COLUMN_NAME = "station_id";
-	public static final String SITE_ID_COLUMN_NAME = "site_id";
-	public static final String EVENT_DATE_COLUMN_NAME = "event_date";
-	public static final String NEMI_URL_COLUMN_NAME = "nemi_url";
-	public static final String PARAMETER_CD_COLUMN_NAME = "parameter_cd";
-	public static final String ACTIVITY_COLUMN_NAME = "activity";
-	public static final String WQPCROSSWALK_COLUMN_NAME = "wqpcrosswalk";
-	public static final String SRSNAME_COLUMN_NAME = "srsname";
-	public static final String PARM_SEQ_GRP_NM_COLUMN_NAME = "parm_seq_grp_nm";
-	public static final String SAMPLE_MEDIA_COLUMN_NAME = "sample_media";
-	public static final String ORGANIZATION_COLUMN_NAME = "organization";
-	public static final String SITE_TYPE_COLUMN_NAME = "site_type";
-	public static final String HUC_COLUMN_NAME = "huc";
-	public static final String GOVERNMENTAL_UNIT_CODE_COLUMN_NAME = "governmental_unit_code";
-	public static final String GEOM_COLUMN_NAME = "geom";
-	public static final String ORGANIZATION_NAME_COLUMN_NAME = "organization_name";
-	public static final String ACTIVITY_ID_COLUMN_NAME = "activity_id";
-	public static final String ACTIVITY_TYPE_CODE_COLUMN_NAME = "activity_type_code";
-	public static final String ACTIVITY_MEDIA_SUBDIV_NAME_COLUMN_NAME = "activity_media_subdiv_name";
-	public static final String ACTIVITY_START_TIME_COLUMN_NAME = "activity_start_time";
-	public static final String ACT_START_TIME_ZONE_COLUMN_NAME = "act_start_time_zone";
-	public static final String ACTIVITY_STOP_DATE_COLUMN_NAME = "activity_stop_date";
-	public static final String ACTIVITY_STOP_TIME_COLUMN_NAME = "activity_stop_time";
-	public static final String ACT_STOP_TIME_ZONE_COLUMN_NAME = "act_stop_time_zone";
-	public static final String ACTIVITY_DEPTH_COLUMN_NAME = "activity_depth";
-	public static final String ACTIVITY_DEPTH_UNIT_COLUMN_NAME = "activity_depth_unit";
-	public static final String ACTIVITY_DEPTH_REF_POINT_COLUMN_NAME = "activity_depth_ref_point";
-	public static final String ACTIVITY_UPPER_DEPTH_COLUMN_NAME = "activity_upper_depth";
-	public static final String ACTIVITY_UPPER_DEPTH_UNIT_COLUMN_NAME = "activity_upper_depth_unit";
-	public static final String ACTIVITY_LOWER_DEPTH_COLUMN_NAME = "activity_lower_depth";
-	public static final String ACTIVITY_LOWER_DEPTH_UNIT_COLUMN_NAME = "activity_lower_depth_unit";
-	public static final String PROJECT_ID_COLUMN_NAME = "project_id";
-	public static final String ACTIVITY_CONDUCTING_ORG_COLUMN_NAME = "activity_conducting_org";
-	public static final String ACTIVITY_COMMENT_COLUMN_NAME = "activity_comment";
-	public static final String SAMPLE_AQFR_NAME_COLUMN_NAME = "sample_aqfr_name";
-	public static final String HYDROLOGIC_CONDITION_NAME_COLUMN_NAME = "hydrologic_condition_name";
-	public static final String HYDROLOGIC_EVENT_NAME_COLUMN_NAME = "hydrologic_event_name";
-	public static final String SAMPLE_COLLECT_METHOD_ID_COLUMN_NAME = "sample_collect_method_id";
-	public static final String SAMPLE_COLLECT_METHOD_CTX_COLUMN_NAME = "sample_collect_method_ctx";
-	public static final String SAMPLE_COLLECT_METHOD_NAME_COLUMN_NAME = "sample_collect_method_name";
-	public static final String SAMPLE_COLLECT_EQUIP_NAME_COLUMN_NAME = "sample_collect_equip_name";
-	public static final String RESULT_ID_COLUMN_NAME = "result_id";
-	public static final String REMARK_CD_COLUMN_NAME = "remark_cd";
-	public static final String PARM_FRAC_TX_COLUMN_NAME = "parm_frac_tx";
-	public static final String FXD_TX_COLUMN_NAME = "fxd_tx";
-	public static final String RESULT_VA_COLUMN_NAME = "result_va";
-	public static final String PARM_UNT_TX_COLUMN_NAME = "parm_unt_tx";
-	public static final String DQI_CD_COLUMN_NAME = "dqi_cd";
-	public static final String PARM_STAT_TX_COLUMN_NAME = "parm_stat_tx";
-	public static final String RESULT_MD_COLUMN_NAME = "result_md";
-	public static final String PARM_WT_TX_COLUMN_NAME = "parm_wt_tx";
-	public static final String PARM_TM_TX_COLUMN_NAME = "parm_tm_tx";
-	public static final String PARM_TEMP_TX_COLUMN_NAME = "parm_temp_tx";
-	public static final String PARM_SIZE_TX_COLUMN_NAME = "parm_size_tx";
-	public static final String LAB_STD_VA_COLUMN_NAME = "lab_std_va";
-	public static final String RESULT_LAB_CM_TX_COLUMN_NAME = "result_lab_cm_tx";
-	public static final String PARM_MEDIUM_TX_COLUMN_NAME = "parm_medium_tx";
-	public static final String SAMPLE_TISSUE_TAXONOMIC_NAME_COLUMN_NAME = "sample_tissue_taxonomic_name";
-	public static final String SAMPLE_TISSUE_ANATOMY_NAME_COLUMN_NAME = "sample_tissue_anatomy_name";
-	public static final String METH_CD_COLUMN_NAME = "meth_cd";
-	public static final String METH_NM_COLUMN_NAME = "meth_nm";
-	public static final String CIT_NM_COLUMN_NAME = "cit_nm";
-	public static final String PROTO_ORG_NM_COLUMN_NAME = "proto_org_nm";
-	public static final String ANL_DT_COLUMN_NAME = "anl_dt";
-	public static final String VAL_QUAL_CD1_NM_COLUMN_NAME = "val_qual_cd1_nm";
-	public static final String VAL_QUAL_CD2_NM_COLUMN_NAME = "val_qual_cd2_nm";
-	public static final String VAL_QUAL_CD3_NM_COLUMN_NAME = "val_qual_cd3_nm";
-	public static final String VAL_QUAL_CD4_NM_COLUMN_NAME = "val_qual_cd4_nm";
-	public static final String VAL_QUAL_CD5_NM_COLUMN_NAME = "val_qual_cd5_nm";
-	public static final String RPT_LEV_VA_COLUMN_NAME = "rpt_lev_va";
-	public static final String RESULT_UNRND_VA_COLUMN_NAME = "result_unrnd_va";
-	public static final String PARM_METH_MULTIPLIER_COLUMN_NAME = "parm_meth_multiplier";
-	public static final String PARM_MULTIPLIER_COLUMN_NAME = "parm_multiplier";
-	public static final String RPT_LEV_CD_COLUMN_NAME = "rpt_lev_cd";
-	public static final String RPT_LEV_NM_COLUMN_NAME = "rpt_lev_nm";
-	public static final String PREP_DT_COLUMN_NAME = "prep_dt";
+	private static final String STATION_ID_COLUMN_NAME = "station_id";
+	private static final String SITE_ID_COLUMN_NAME = "site_id";
+	private static final String EVENT_DATE_COLUMN_NAME = "event_date";
+	private static final String NEMI_URL_COLUMN_NAME = "nemi_url";
+	private static final String PARAMETER_CD_COLUMN_NAME = "parameter_cd";
+	private static final String ACTIVITY_COLUMN_NAME = "activity";
+	private static final String WQPCROSSWALK_COLUMN_NAME = "wqpcrosswalk";
+	private static final String SRSNAME_COLUMN_NAME = "srsname";
+	private static final String PARM_SEQ_GRP_NM_COLUMN_NAME = "parm_seq_grp_nm";
+	private static final String SAMPLE_MEDIA_COLUMN_NAME = "sample_media";
+	private static final String ORGANIZATION_COLUMN_NAME = "organization";
+	private static final String SITE_TYPE_COLUMN_NAME = "site_type";
+	private static final String HUC_COLUMN_NAME = "huc";
+	private static final String GOVERNMENTAL_UNIT_CODE_COLUMN_NAME = "governmental_unit_code";
+	private static final String GEOM_COLUMN_NAME = "geom";
+	private static final String ORGANIZATION_NAME_COLUMN_NAME = "organization_name";
+	private static final String ACTIVITY_ID_COLUMN_NAME = "activity_id";
+	private static final String ACTIVITY_TYPE_CODE_COLUMN_NAME = "activity_type_code";
+	private static final String ACTIVITY_MEDIA_SUBDIV_NAME_COLUMN_NAME = "activity_media_subdiv_name";
+	private static final String ACTIVITY_START_TIME_COLUMN_NAME = "activity_start_time";
+	private static final String ACT_START_TIME_ZONE_COLUMN_NAME = "act_start_time_zone";
+	private static final String ACTIVITY_STOP_DATE_COLUMN_NAME = "activity_stop_date";
+	private static final String ACTIVITY_STOP_TIME_COLUMN_NAME = "activity_stop_time";
+	private static final String ACT_STOP_TIME_ZONE_COLUMN_NAME = "act_stop_time_zone";
+	private static final String ACTIVITY_DEPTH_COLUMN_NAME = "activity_depth";
+	private static final String ACTIVITY_DEPTH_UNIT_COLUMN_NAME = "activity_depth_unit";
+	private static final String ACTIVITY_DEPTH_REF_POINT_COLUMN_NAME = "activity_depth_ref_point";
+	private static final String ACTIVITY_UPPER_DEPTH_COLUMN_NAME = "activity_upper_depth";
+	private static final String ACTIVITY_UPPER_DEPTH_UNIT_COLUMN_NAME = "activity_upper_depth_unit";
+	private static final String ACTIVITY_LOWER_DEPTH_COLUMN_NAME = "activity_lower_depth";
+	private static final String ACTIVITY_LOWER_DEPTH_UNIT_COLUMN_NAME = "activity_lower_depth_unit";
+	private static final String PROJECT_ID_COLUMN_NAME = "project_id";
+	private static final String ACTIVITY_CONDUCTING_ORG_COLUMN_NAME = "activity_conducting_org";
+	private static final String ACTIVITY_COMMENT_COLUMN_NAME = "activity_comment";
+	private static final String SAMPLE_AQFR_NAME_COLUMN_NAME = "sample_aqfr_name";
+	private static final String HYDROLOGIC_CONDITION_NAME_COLUMN_NAME = "hydrologic_condition_name";
+	private static final String HYDROLOGIC_EVENT_NAME_COLUMN_NAME = "hydrologic_event_name";
+	private static final String SAMPLE_COLLECT_METHOD_ID_COLUMN_NAME = "sample_collect_method_id";
+	private static final String SAMPLE_COLLECT_METHOD_CTX_COLUMN_NAME = "sample_collect_method_ctx";
+	private static final String SAMPLE_COLLECT_METHOD_NAME_COLUMN_NAME = "sample_collect_method_name";
+	private static final String SAMPLE_COLLECT_EQUIP_NAME_COLUMN_NAME = "sample_collect_equip_name";
+	private static final String RESULT_ID_COLUMN_NAME = "result_id";
+	private static final String REMARK_CD_COLUMN_NAME = "remark_cd";
+	private static final String PARM_FRAC_TX_COLUMN_NAME = "parm_frac_tx";
+	private static final String FXD_TX_COLUMN_NAME = "fxd_tx";
+	private static final String RESULT_VA_COLUMN_NAME = "result_va";
+	private static final String PARM_UNT_TX_COLUMN_NAME = "parm_unt_tx";
+	private static final String DQI_CD_COLUMN_NAME = "dqi_cd";
+	private static final String PARM_STAT_TX_COLUMN_NAME = "parm_stat_tx";
+	private static final String RESULT_MD_COLUMN_NAME = "result_md";
+	private static final String PARM_WT_TX_COLUMN_NAME = "parm_wt_tx";
+	private static final String PARM_TM_TX_COLUMN_NAME = "parm_tm_tx";
+	private static final String PARM_TEMP_TX_COLUMN_NAME = "parm_temp_tx";
+	private static final String PARM_SIZE_TX_COLUMN_NAME = "parm_size_tx";
+	private static final String LAB_STD_VA_COLUMN_NAME = "lab_std_va";
+	private static final String RESULT_LAB_CM_TX_COLUMN_NAME = "result_lab_cm_tx";
+	private static final String PARM_MEDIUM_TX_COLUMN_NAME = "parm_medium_tx";
+	private static final String SAMPLE_TISSUE_TAXONOMIC_NAME_COLUMN_NAME = "sample_tissue_taxonomic_name";
+	private static final String SAMPLE_TISSUE_ANATOMY_NAME_COLUMN_NAME = "sample_tissue_anatomy_name";
+	private static final String METH_CD_COLUMN_NAME = "meth_cd";
+	private static final String METH_NM_COLUMN_NAME = "meth_nm";
+	private static final String CIT_NM_COLUMN_NAME = "cit_nm";
+	private static final String PROTO_ORG_NM_COLUMN_NAME = "proto_org_nm";
+	private static final String ANL_DT_COLUMN_NAME = "anl_dt";
+	private static final String VAL_QUAL_CD1_NM_COLUMN_NAME = "val_qual_cd1_nm";
+	private static final String VAL_QUAL_CD2_NM_COLUMN_NAME = "val_qual_cd2_nm";
+	private static final String VAL_QUAL_CD3_NM_COLUMN_NAME = "val_qual_cd3_nm";
+	private static final String VAL_QUAL_CD4_NM_COLUMN_NAME = "val_qual_cd4_nm";
+	private static final String VAL_QUAL_CD5_NM_COLUMN_NAME = "val_qual_cd5_nm";
+	private static final String RPT_LEV_VA_COLUMN_NAME = "rpt_lev_va";
+	private static final String RESULT_UNRND_VA_COLUMN_NAME = "result_unrnd_va";
+	private static final String PARM_METH_MULTIPLIER_COLUMN_NAME = "parm_meth_multiplier";
+	private static final String PARM_MULTIPLIER_COLUMN_NAME = "parm_multiplier";
+	private static final String RPT_LEV_CD_COLUMN_NAME = "rpt_lev_cd";
+	private static final String RPT_LEV_NM_COLUMN_NAME = "rpt_lev_nm";
+	private static final String PREP_DT_COLUMN_NAME = "prep_dt";
 
 	@Override
 	public NwisResult mapRow(ResultSet rs, int rowNum) throws SQLException {
